Move 2D visualization into its own module
